requirements/base.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile --output-file=base.txt base.in
#
amqp==5.1.1
    # via kombu
asgiref==3.5.2
    # via
    #   django
    #   djoser
asttokens==2.1.0
    # via stack-data
backcall==0.2.0
    # via ipython
billiard==3.6.4.0
    # via celery
celery==5.2.7
    # via -r base.in
certifi==2022.9.24
    # via requests
cffi==1.15.1
    # via cryptography
charset-normalizer==2.1.1
    # via requests
click==8.1.3
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
click-didyoumean==0.3.0
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.2.0
    # via celery
coreapi==2.3.3
    # via
    #   djoser
    #   drf-yasg
coreschema==0.0.4
    # via
    #   coreapi
    #   drf-yasg
cryptography==38.0.3
    # via social-auth-core
decorator==5.1.1
    # via ipython
defusedxml==0.7.1
    # via
    #   python3-openid
    #   social-auth-core
django==4.1.3
    # via
    #   -r base.in
    #   django-admin-display
    #   django-dbbackup
    #   django-debug-toolbar
    #   django-extensions
    #   django-filter
    #   django-jazzmin
    #   djangorestframework
    #   djangorestframework-simplejwt
    #   drf-yasg
django-admin-display==1.3.0
    # via -r base.in
django-dbbackup==4.0.2
    # via -r base.in
django-debug-toolbar==3.7.0
    # via -r base.in
django-extensions==3.2.1
    # via -r base.in
django-filter==22.1
    # via -r base.in
django-jazzmin==2.6.0
    # via -r base.in
django-templated-mail==1.1.1
    # via djoser
django-webpack-loader==1.7.0
    # via -r base.in
djangorestframework==3.14.0
    # via
    #   -r base.in
    #   djangorestframework-simplejwt
    #   drf-yasg
djangorestframework-simplejwt==4.8.0
    # via
    #   -r base.in
    #   djoser
djoser==2.1.0
    # via -r base.in
drf-flex-fields==1.0.0
    # via -r base.in
drf-yasg==1.21.4
    # via -r base.in
environs==9.5.0
    # via -r base.in
executing==1.2.0
    # via stack-data
idna==3.4
    # via requests
inflection==0.5.1
    # via drf-yasg
ipython==8.6.0
    # via -r base.in
itypes==1.2.0
    # via coreapi
jedi==0.18.1
    # via ipython
jinja2==3.1.2
    # via coreschema
kombu==5.2.4
    # via celery
markupsafe==2.1.1
    # via jinja2
marshmallow==3.19.0
    # via environs
matplotlib-inline==0.1.6
    # via ipython
oauthlib==3.2.2
    # via
    #   requests-oauthlib
    #   social-auth-core
packaging==21.3
    # via
    #   drf-yasg
    #   marshmallow
parso==0.8.3
    # via jedi
pexpect==4.8.0
    # via ipython
pickleshare==0.7.5
    # via ipython
prompt-toolkit==3.0.32
    # via
    #   click-repl
    #   ipython
psycopg2-binary==2.9.5
    # via -r base.in
ptyprocess==0.7.0
    # via pexpect
pure-eval==0.2.2
    # via stack-data
pycparser==2.21
    # via cffi
pygments==2.13.0
    # via ipython
pyjwt==2.6.0
    # via
    #   djangorestframework-simplejwt
    #   social-auth-core
pyparsing==3.0.9
    # via packaging
python-dotenv==0.21.0
    # via environs
python3-openid==3.2.0
    # via social-auth-core
pytz==2022.6
    # via
    #   celery
    #   django-dbbackup
    #   djangorestframework
    #   drf-yasg
requests==2.28.1
    # via
    #   coreapi
    #   requests-oauthlib
    #   social-auth-core
requests-oauthlib==1.3.1
    # via social-auth-core
ruamel-yaml==0.17.21
    # via drf-yasg
ruamel-yaml-clib==0.2.7
    # via ruamel-yaml
six==1.16.0
    # via
    #   asttokens
    #   click-repl
    #   social-auth-app-django
social-auth-app-django==4.0.0
    # via djoser
social-auth-core==4.3.0
    # via social-auth-app-django
sqlparse==0.4.3
    # via
    #   django
    #   django-debug-toolbar
stack-data==0.6.1
    # via ipython
traitlets==5.5.0
    # via
    #   ipython
    #   matplotlib-inline
uritemplate==4.1.1
    # via
    #   coreapi
    #   drf-yasg
urllib3==1.26.12
    # via requests
vine==5.0.0
    # via
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.5
    # via prompt-toolkit
